Just try Postman for Users and Account (Post one first and then try to get, put and delete)

+++ src/main/java/com/cpo/bank/controller/AccountController.java
@@ -22,7 +22,7 @@ public class AccountController {
 	
 	// get Account by ID
 	@GetMapping("/account/{accountID}")
-	private Account getAccount(@PathVariable("accountID") int accountID) {
+	private Account getAccount(@PathVariable("accountID")long accountID) {
 		return accountService.getAccountById(accountID);
 	}
